refactor(template-golang-lib): restructure project layout and improve documentation
- Move example_test.go to example directory- Update package name in example_test.go
- Add godoc support with new go-doc.mk file
- Improve help messages in Makefile utilities
- Update Makefile to include godoc configuration
