+++ Makefile
@@ -35,6 +35,11 @@ endif
 ENV_ROOT_TEST_MAX_TIME :=1m
 ## go test go-test.mk end
 
+## godoc start
+ENV_GO_GODOC_PORT_NUMBER=36060
+include z-MakefileUtils/go-doc.mk
+## godoc end
+
 ## clean args start
 ENV_ROOT_BUILD_PATH =build
 ENV_ROOT_LOG_PATH =logs/
