+++ example/example_test.go
@@ -1,15 +1,14 @@
-package template_golang_lib
+package example
 
 import (
 	"fmt"
-	"github.com/bridgewwater/template-golang-lib/example"
 )
 
 // To use template-golang-lib lib package, just
 // the SetOutput function when your application starts.
 func Example() {
 	// Basic usage of this lib
-	fooOut := example.Foo(1, 2)
+	fooOut := Foo(1, 2)
 
 	fmt.Printf("fooOut %v\n", fooOut)
 	// Output:
@@ -20,7 +19,7 @@ func Example() {
 // When parsing, the first letter of the Example_ content will be automatically capitalized.
 func Example_bar() {
 	// For some usage of example.Foo
-	barOut := example.Foo(3, 4)
+	barOut := Foo(3, 4)
 
 	fmt.Printf("barOut %v", barOut)
 	// Output:
