+++ example/foo.go
@@ -1,5 +1,7 @@
 package example
 
+// Foo
+// example method for show
 func Foo(one, two int) int {
 	return one + two
 }
